backend/pyproject.toml
[project]
name = "libro-digital"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "authlib==1.6.0",
    "bcrypt==4.3.0",
    "certifi==2025.6.15",
    "cffi==1.17.1",
    "charset-normalizer==3.4.2",
    "click==8.2.1",
    "colorama==0.4.6",
    "cryptography==45.0.4",
    "ecdsa==0.19.1",
    "fastapi[standard]==0.115.12",
    "greenlet==3.2.3",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "idna==3.10",
    "mercadopago==2.3.0",
    "passlib==1.7.4",
    "pillow>=11.3.0",
    "psycopg2-binary==2.9.10",
    "pyasn1==0.6.1",
    "pycparser==2.22",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "python-dotenv==1.1.0",
    "python-jose==3.5.0",
    "python-multipart==0.0.20",
    "requests==2.32.4",
    "resend==2.10.0",
    "rsa==4.9.1",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.41",
    "sqlmodel==0.0.24",
    "starlette==0.46.2",
    "typing-extensions==4.14.0",
    "typing-inspection==0.4.1",
    "urllib3==2.4.0",
    "uvicorn==0.34.3",
]

[tool.fastapi]
app = "app.main:app"
